Extract FileOperations into separate object
